requirements.txt
beautifulsoup4==4.13.5
blinker==1.9.0
click==8.2.1
colorama==0.4.6
Flask==3.1.2
html5lib==1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
pillow==11.3.0
six==1.17.0
soupsieve==2.8
typing_extensions==4.15.0
wcag-contrast-ratio==0.9
webencodings==0.5.1
Werkzeug==3.1.3
